+++ lib/src/navbar_item_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'responsiveness.dart';
 
 class NavbarItemWidget extends StatelessWidget {
   final Widget icon;
@@ -53,14 +53,12 @@ class NavbarItemWidget extends StatelessWidget {
             FittedBox(
               fit: BoxFit.contain,
               alignment: Alignment.center,
-
               child: Text(
                 label,
                 style: TextStyle(
                   color: isSelected ? selectedColor : unselectedColor,
-                  fontSize: isSelected
-                      ? selectedFontSize.sp
-                      : unselectedFontSize.sp,
+                  fontSize:
+                      isSelected ? selectedFontSize.sp : unselectedFontSize.sp,
                   fontWeight: isSelected ? FontWeight.bold : FontWeight.normal,
                 ),
               ),